requirements.txt
Django
Pillow
django-tastypie
django-spa
djangorestframework-jwt
djangorestframework==3.11.2
whitenoise
drf-yasg
packaging